tabibi_dz/requirements.txt
annotated-types
asgiref
asttokens
cachetools
certifi
python-dotenv
charset-normalizer
colorama
comm
debugpy
decorator
dj-database-url
Django
django-allauth
django-cors-headers
django-heroku
djangorestframework
djangorestframework-simplejwt
executing
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
gunicorn
httplib2
idna
ipykernel
ipython
jedi
jupyter_client
jupyter_core
matplotlib-inline
psycopg2 
nest-asyncio
packaging
parso
platformdirs
prompt_toolkit
proto-plus
protobuf
psutil
psycopg2
pure-eval
pyasn1
pyasn1_modules

pydantic_core

PyJWT
pyparsing
python-dateutil

pyzmq
requests
rsa
six
sqlparse
stack-data
tornado
tqdm
traitlets
typing_extensions
tzdata
uritemplate
urllib3
wcwidth
whitenoise
mysqlclient>=2.1.0  # MySQL adapter for Django
